Refactoring secondary form add education&Projects form
