+++ src/components/EducationForm/styles.js
@@ -1,0 +1,8 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+export const useStyles = makeStyles(() => ({
+  educationContainer: {
+    padding: '10px',
+    cursor: 'pointer',
+  },
+}));
