+++ src/components/ProjectsForm/styles.js
@@ -1,0 +1,14 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+export const useStyles = makeStyles(() => ({
+  input: {
+    width: '100%',
+  },
+  projTitle: {
+    fontStyle: 'italic',
+    fontWeight: 'bold',
+  },
+  removeButton: {
+    width: '250px',
+  },
+}));
