+++ src/components/SecondarySectionPartForm/SecondarySectionPartForm.jsx
@@ -1,24 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {
-  Typography,
-  InputAdornment,
-  IconButton,
-  Button,
-  TextField,
-  Grid,
-  Tooltip,
-  InputLabel,
-} from '@material-ui/core';
+import { Typography, InputAdornment, IconButton, Grid, InputLabel } from '@material-ui/core';
 import RemoveCircleOutlineIcon from '@material-ui/icons/RemoveCircleOutline';
-import DeleteIcon from '@material-ui/icons/Delete';
 import AddIcon from '@material-ui/icons/Add';
 import { useStyles } from './styles';
 import Card from '@material-ui/core/Card';
 import OutlinedInput from '@material-ui/core/OutlinedInput';
 import { handleInput } from '../../services/HandleInput';
 
-import { disablebAddField } from '../../services/validationAddField';
+import ProjectsForm from '../ProjectsForm/ProjectsForm';
 
 export default function SecondarySectionPartForm({
   sectionsField,
@@ -38,14 +28,7 @@ export default function SecondarySectionPartForm({
   const isObject = arg => {
     return !!arg && arg.constructor === Object;
   };
-  const splitValue = proj => {
-    const keyProject = Object.keys(proj);
-    if (proj[keyProject]['Team']) {
-      const teamInfo = proj[keyProject]['Team'];
-      const getNumber = teamInfo.split(' ')[0];
-      return getNumber;
-    }
-  };
+
   return (
     <>
       {Object.entries(sectionsField).map(([key, value]) =>
@@ -61,180 +44,17 @@ export default function SecondarySectionPartForm({
               {Object.entries(value).map(([label, defaultValue]) => (
                 <React.Fragment key={label}>
                   {label === '$projects' ? (
-                    <React.Fragment key={`${defaultValue}-grid`}>
-                      {sectionsField['SIGNIFICANT PROJECTS'].$projects.map((proj, indexProj) => (
-                        <React.Fragment key={`${Object.keys(proj)}-rpoj-fragment`}>
-                          <Grid item sm={12} xs={12} key={`${Object.keys(proj)}-grid`}>
-                            <Typography
-                              key={Object.keys(proj)}
-                              fontStyle="italic"
-                              className={classes.projTitle}
-                              variant="subtitle2"
-                              gutterBottom
-                            >
-                              {Object.keys(proj)}
-                            </Typography>
-                          </Grid>
-                          <Grid item sm={12} xs={12}>
-                            <TextField
-                              className={classes.input}
-                              key={`${Object.keys(proj)}-name`}
-                              fullWidth
-                              variant="outlined"
-                              label="Name"
-                              defaultValue={Object.keys(proj)}
-                              onBlur={e =>
-                                changeProjectName(e.target.value, indexProj, Object.keys(proj))
-                              }
-                            />
-                          </Grid>
-                          <Grid item xs={12}>
-                            <TextField
-                              className={classes.input}
-                              fullWidth
-                              key={proj[Object.keys(proj)]['$description']}
-                              label="Description"
-                              variant="outlined"
-                              multiline
-                              rows={4}
-                              defaultValue={proj[Object.keys(proj)]['$description']}
-                              onBlur={e =>
-                                handleInput(
-                                  setGlobalError,
-                                  e.target.value,
-                                  setSingleFieldProject(
-                                    e.target.value,
-                                    proj,
-                                    '$description',
-                                    indexProj,
-                                  ),
-                                )
-                              }
-                            />
-                          </Grid>
-                          <>
-                            {proj[Object.keys(proj)]['Team'] ? (
-                              <Grid item sm={6} xs={12}>
-                                <TextField
-                                  type="number"
-                                  fullWidth
-                                  key={proj[Object.keys(proj)]['Team']}
-                                  className={classes.input}
-                                  label="Team"
-                                  variant="outlined"
-                                  defaultValue={splitValue(proj)}
-                                  onBlur={e =>
-                                    setSingleFieldProject(+e.target.value, proj, 'Team', indexProj)
-                                  }
-                                />
-                              </Grid>
-                            ) : null}
-                          </>
-                          <Grid item sm={proj[Object.keys(proj)]['Team'] ? 6 : 12} xs={12}>
-                            <TextField
-                              className={classes.input}
-                              key={proj[Object.keys(proj)]['Skills']}
-                              fullWidth
-                              variant="outlined"
-                              label="Skills"
-                              defaultValue={proj[Object.keys(proj)]['Skills']}
-                              onBlur={e =>
-                                handleInput(
-                                  setGlobalError,
-                                  e.target.value,
-                                  setSingleFieldProject(e.target.value, proj, 'Skills', indexProj),
-                                )
-                              }
-                            />
-                          </Grid>
-                          <Grid item xs={12}>
-                            <TextField
-                              className={classes.input}
-                              fullWidth
-                              key={proj[Object.keys(proj)]['Role']}
-                              label="Role"
-                              variant="outlined"
-                              defaultValue={proj[Object.keys(proj)]['Role']}
-                              onBlur={e =>
-                                handleInput(
-                                  setGlobalError,
-                                  e.target.value,
-                                  setSingleFieldProject(e.target.value, proj, 'Role', indexProj),
-                                )
-                              }
-                            />
-                          </Grid>
-                          <Grid item sm={12} xs={12}>
-                            <Typography>
-                              Responsibilities
-                              <Tooltip
-                                element={'span'}
-                                placement="right"
-                                title={
-                                  disablebAddField(proj[Object.keys(proj)]['Responsibilities']) ? (
-                                    <span style={{ fontSize: '22px' }}>
-                                      Please fill all input fields
-                                    </span>
-                                  ) : (
-                                    ''
-                                  )
-                                }
-                              >
-                                <span>
-                                  <IconButton
-                                    disabled={disablebAddField(
-                                      proj[Object.keys(proj)]['Responsibilities'],
-                                    )}
-                                    variant="contained"
-                                    onClick={() => addFieldResponsibility(proj, indexProj)}
-                                  >
-                                    <AddIcon />
-                                  </IconButton>
-                                </span>
-                              </Tooltip>
-                            </Typography>
-                          </Grid>
-                          {proj[Object.keys(proj)]['Responsibilities'].map((res, index) => (
-                            <Grid item sm={12} xs={12} key={res}>
-                              <OutlinedInput
-                                className={classes.input}
-                                fullWidth
-                                defaultValue={res}
-                                onBlur={e =>
-                                  handleInput(
-                                    setGlobalError,
-                                    e.target.value,
-                                    setValueResponsibility(e.target.value, proj, index, indexProj),
-                                  )
-                                }
-                                endAdornment={
-                                  <InputAdornment position="end">
-                                    <IconButton
-                                      variant="contained"
-                                      color="secondary"
-                                      onClick={() =>
-                                        removeFieldResponsibility(proj, index, indexProj)
-                                      }
-                                    >
-                                      <RemoveCircleOutlineIcon />
-                                    </IconButton>
-                                  </InputAdornment>
-                                }
-                              />
-                            </Grid>
-                          ))}
-                          <Button
-                            title={`Remove project ${Object.keys(proj)}`}
-                            variant="contained"
-                            color="secondary"
-                            onClick={() => removeProject(indexProj)}
-                            endIcon={<DeleteIcon />}
-                          >
-                            {`Remove project ${Object.keys(proj)}`}
-                          </Button>
-                        </React.Fragment>
-                      ))}
-                    </React.Fragment>
+                    <ProjectsForm
+                      defaultValue={defaultValue}
+                      sectionsField={sectionsField}
+                      changeProjectName={changeProjectName}
+                      setGlobalError={setGlobalError}
+                      setSingleFieldProject={setSingleFieldProject}
+                      addFieldResponsibility={addFieldResponsibility}
+                      setValueResponsibility={setValueResponsibility}
+                      removeFieldResponsibility={removeFieldResponsibility}
+                      removeProject={removeProject}
+                    />
                   ) : (
                     <Grid item sm={6} xs={12} key={`${defaultValue}-grid`}>
                       <InputLabel htmlFor={`${label}-component-outlined`}>
